Added opal datasource to list of available databases. Made some databases immutable. Added custom dialect resolver to Hibernate to ensure InnoDB and custom HSQLDB dialect.

+++ opal-gwt-client/src/main/java/org/obiba/opal/web/gwt/app/client/administration/database/presenter/DatabaseAdministrationPresenter.java
@@ -80,15 +80,11 @@ public class DatabaseAdministrationPresenter extends ItemAdministrationPresenter
 
   }
 
-  private static final String DATABASES_RESOURCE = "/jdbc/databases";
-
-  private static final String DATABASE_RESOURCE = "/jdbc/database/";
-
   private final Provider<DatabasePresenter> jdbcDataSourcePresenter;
 
   private final AuthorizationPresenter authorizationPresenter;
 
-  private final ResourceDataProvider<JdbcDataSourceDto> resourceDataProvider = new ResourceDataProvider<JdbcDataSourceDto>(DATABASES_RESOURCE);
+  private final ResourceDataProvider<JdbcDataSourceDto> resourceDataProvider = new ResourceDataProvider<JdbcDataSourceDto>(Resources.databases());
 
   private Command confirmedCommand;
 
@@ -102,7 +98,7 @@ public class DatabaseAdministrationPresenter extends ItemAdministrationPresenter
   @ProxyEvent
   @Override
   public void onAdministrationPermissionRequest(RequestAdministrationPermissionEvent event) {
-    ResourceAuthorizationRequestBuilderFactory.newBuilder().forResource(DATABASES_RESOURCE).post().authorize(event.getHasAuthorization()).send();
+    ResourceAuthorizationRequestBuilderFactory.newBuilder().forResource(Resources.databases()).post().authorize(event.getHasAuthorization()).send();
   }
 
   @Override
@@ -123,7 +119,7 @@ public class DatabaseAdministrationPresenter extends ItemAdministrationPresenter
 
   @Override
   public void authorize(HasAuthorization authorizer) {
-    ResourceAuthorizationRequestBuilderFactory.newBuilder().forResource(DATABASES_RESOURCE).post().authorize(authorizer).send();
+    ResourceAuthorizationRequestBuilderFactory.newBuilder().forResource(Resources.databases()).post().authorize(authorizer).send();
   }
 
   @Override
@@ -157,14 +153,14 @@ public class DatabaseAdministrationPresenter extends ItemAdministrationPresenter
 
       @Override
       public void doAction(final JdbcDataSourceDto object, String actionName) {
-        if(actionName.equalsIgnoreCase(DELETE_ACTION)) {
+        if(object.getEditable() && actionName.equalsIgnoreCase(DELETE_ACTION)) {
           getEventBus().fireEvent(new ConfirmationRequiredEvent(confirmedCommand = new Command() {
             @Override
             public void execute() {
               deleteDatabase(object);
             }
           }, "deleteDatabase", "confirmDeleteDatabase"));
-        } else if(actionName.equalsIgnoreCase(EDIT_ACTION)) {
+        } else if(object.getEditable() && actionName.equalsIgnoreCase(EDIT_ACTION)) {
           DatabasePresenter dialog = jdbcDataSourcePresenter.get();
           dialog.updateDatabase(object);
           addToPopupSlot(dialog);
@@ -183,7 +179,7 @@ public class DatabaseAdministrationPresenter extends ItemAdministrationPresenter
 
           };
           ResourceRequestBuilderFactory.<JsArray<JdbcDataSourceDto>> newBuilder()//
-          .forResource(databaseResource(object.getName(), "connections")).accept("application/json")//
+          .forResource(Resources.database(object.getName(), "connections")).accept("application/json")//
           .withCallback(200, callback).withCallback(503, callback).post().send();
         }
       }
@@ -202,11 +198,11 @@ public class DatabaseAdministrationPresenter extends ItemAdministrationPresenter
     }));
 
     resourceDataProvider.addDataDisplay(getView().getDatabaseTable());
-    authorizationPresenter.setAclRequest("databases", AclRequest.newBuilder("Administrate", DATABASES_RESOURCE, "*:POST/*"));
+    authorizationPresenter.setAclRequest("databases", AclRequest.newBuilder("Administrate", Resources.databases(), "*:POST/*"));
   }
 
   private void deleteDatabase(JdbcDataSourceDto database) {
-    ResourceRequestBuilderFactory.<JsArray<JdbcDataSourceDto>> newBuilder().forResource(databaseResource(database.getName())).withCallback(200, new ResponseCodeCallback() {
+    ResourceRequestBuilderFactory.<JsArray<JdbcDataSourceDto>> newBuilder().forResource(Resources.database(database.getName())).withCallback(200, new ResponseCodeCallback() {
 
       @Override
       public void onResponseCode(Request request, Response response) {
@@ -218,16 +214,6 @@ public class DatabaseAdministrationPresenter extends ItemAdministrationPresenter
 
   private void refresh() {
     getView().getDatabaseTable().setVisibleRangeAndClearData(new Range(0, 10), true);
-  }
-
-  private String databaseResource(String databaseName, String... subResources) {
-    String resource = DATABASE_RESOURCE + databaseName;
-    if(subResources != null) {
-      for(String s : subResources) {
-        resource += "/" + s;
-      }
-    }
-    return resource;
   }
 
   private final class PermissionsUpdate implements HasAuthorization {
